add dynamic routing for the events route
