+++ public/index.html
@@ -7,6 +7,6 @@
     </head>
     <body>
       <div id='app'></div>
-      <script src='./bundle.js'></script>
+      <script src='/bundle.js'></script>
     </body>
 </html>
